add colno/end_colno and end_linno properties to JSON5Token. restructure JSONObject to store keys and values separately
This change removes `KeyValuePair` as a Node object.
Instead, the underlying data model will store a list of keys and list of values in separate attributes.
A new attribute for `JSONOBject` `key_value_pairs` which will be a named tuple (typing.NamedTuple) that
provides the key and value attributes.

This change brings more congruity with how the stdlib `ast` module handles Python dictionaries, to which
Python users may be more accustomed. It also maintains, by interface, the same structure for JSON5
members (https://spec.json5.org/#prod-JSON5MemberList) with the new key_value_pairs attribute (although it itself is not a Node).